bare minimum changes to compile (partly) on MSVC
Changes:
- made it possible to override ANSILOVE_EXTERN
  with something that's not GNU specific
- re-implemented strndup in compat/

This makes it possible to compile the core functionality
of libansilove (specifically, ansilove_ansi() and friends)
with MSVC, if some implementation of GD is available.
The _init(), _loadfile() and _clean() functions don't work
yet, as they are centered around POSIX-exclusive APIs.
This functionality is easily replaced by a few lines of
glue code though.